Update README.md
The series revolves around street racing on Tokyo's Bayshore Route (湾岸, Wangan), as well as other roads on the Shuto Expressway network. The story begins with high school student Akio Asakura encountering a black Porsche 911 Turbo nicknamed "Blackbird" on the Wangan. Akio attempts to give chase, but is unable to keep up with the Blackbird, which is driven by a medical doctor named Tatsuya Shima.